Add doSetDisplayModeTest
To test method of changing current display mode to a specific mode.
It seems that you can only set modes to those that have kDisplayModeValidFlag; kDisplayModeValidForMirroringFlag is not sufficient.
Discovered while working on downsampling patches in 10.5.8. Patches for 10.4.11 and 10.6.8 were also created.
See https://forums.macrumors.com/threads/civilization-iii-on-clamshell-ibook-help-write-history.2388728/post-32149906